add: IdentityUser replaced with User, DTOs changed, frontend changed, security system partially changed

+++ ToDo.WebAPI/Controllers/ToDoItemController.cs
@@ -1,8 +1,10 @@
 ﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using ToDo.DomainModel.Classes;
+using ToDo.DomainModel.Models;
+using ToDo.Services.DTOs;
 using ToDo.Services.Interfaces;
 
 namespace ToDo.WebAPI.Controllers
@@ -16,25 +18,20 @@ namespace ToDo.WebAPI.Controllers
     public class ToDoItemController : ControllerBase
     {
         private readonly IToDoItemService toDoItemService;
+        private readonly IToDoListService toDoListService;
+        private readonly IHttpContextService httpContextService;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ToDoItemController"/> class.
         /// </summary>
-        /// <param name="toDoItemService">Service for <see cref="ToDoList"/>.</param>
-        public ToDoItemController(IToDoItemService toDoItemService)
+        /// <param name="toDoItemService">Service for <see cref="ToDoItem"/>.</param>
+        /// <param name="httpContextService">Service for http context.</param>
+        /// <param name="toDoListService">Service for <see cref="ToDoList"/>.</param>
+        public ToDoItemController(IToDoItemService toDoItemService, IHttpContextService httpContextService, IToDoListService toDoListService)
         {
             this.toDoItemService = toDoItemService;
-        }
-
-        /// <summary>
-        /// Handles request for getting all <see cref="ToDoItem"/>.
-        /// </summary>
-        /// <returns>Http response with the list of all <see cref="ToDoItem"/>.</returns>
-        [HttpGet]
-        public ActionResult GetAllItems()
-        {
-            var items = this.toDoItemService.GetAllItems();
-            return this.Ok(items);
+            this.httpContextService = httpContextService;
+            this.toDoListService = toDoListService;
         }
 
         /// <summary>
@@ -46,6 +43,11 @@ namespace ToDo.WebAPI.Controllers
         [Route("list/{id:int}")]
         public async Task<IActionResult> GetItemsByListID(int id)
         {
+            if ((await this.toDoListService.GetListByID(id)).UserID != this.httpContextService.GetIdByContextUser(this.HttpContext.User))
+            {
+                return this.BadRequest("This list does not belong to you.");
+            }
+
             var listItems = await this.toDoItemService.GetItemsByListID(id);
             return this.Ok(listItems);
         }
@@ -58,8 +60,9 @@ namespace ToDo.WebAPI.Controllers
         [Route("today")]
         public IActionResult GetToday()
         {
+            int userID = this.httpContextService.GetIdByContextUser(this.HttpContext.User);
             var todayItems = this.toDoItemService.GetItemsByDate(DateTime.Now.Date);
-            return this.Ok(todayItems);
+            return this.Ok(todayItems.Where(i => i.TodoList.UserID == userID));
         }
 
         /// <summary>
@@ -70,8 +73,9 @@ namespace ToDo.WebAPI.Controllers
         [Route("remind")]
         public ActionResult GetReminded()
         {
+            int userID = this.httpContextService.GetIdByContextUser(this.HttpContext.User);
             var remindedItems = this.toDoItemService.GetItemsForReminder();
-            return this.Ok(remindedItems);
+            return this.Ok(remindedItems.Where(i => i.TodoList.UserID == userID));
         }
 
         /// <summary>
@@ -82,8 +86,9 @@ namespace ToDo.WebAPI.Controllers
         [Route("starred")]
         public ActionResult GetStarred()
         {
+            int userID = this.httpContextService.GetIdByContextUser(this.HttpContext.User);
             var starredItems = this.toDoItemService.GetItemsByPriority(Priority.Top);
-            return this.Ok(starredItems);
+            return this.Ok(starredItems.Where(i => i.TodoList.UserID == userID));
         }
 
         /// <summary>
@@ -101,6 +106,11 @@ namespace ToDo.WebAPI.Controllers
                 return this.NotFound();
             }
 
+            if (item.TodoList.UserID != this.httpContextService.GetIdByContextUser(this.HttpContext.User))
+            {
+                return this.BadRequest("This list does not belong to you.");
+            }
+
             return this.Ok(item);
         }
 
@@ -110,8 +120,13 @@ namespace ToDo.WebAPI.Controllers
         /// <param name="item">Updated <see cref="ToDoItem"/>.</param>
         /// <returns>BadRequest response if <see cref="ToDoItem"/> is empty or <see cref="ToDoItem.Id"/> in query and <see cref="ToDoItem.Id"/> in <see cref="ToDoItem"/> (or <see cref="ToDoItem.Id"/> in found item and <see cref="ToDoItem"/> in sent <see cref="ToDoItem"/>) are different, NotFound response if there is no such <see cref="ToDoItem"/>, Ok response with the updated <see cref="ToDoItem"/>.</returns>
         [HttpPut]
-        public async Task<IActionResult> UpdateItem([FromBody] ToDoItem item)
+        public async Task<IActionResult> UpdateItem([FromBody] ToDoItemUpdate item)
         {
+            if ((await this.toDoListService.GetListByID(item.ToDoListID)).UserID != this.httpContextService.GetIdByContextUser(this.HttpContext.User))
+            {
+                return this.BadRequest("This list does not belong to you.");
+            }
+
             try
             {
                 var updatedItem = await this.toDoItemService.UpdateItem(item);
@@ -132,13 +147,18 @@ namespace ToDo.WebAPI.Controllers
         }
 
         /// <summary>
-        /// Handles request for inserting <see cref="ToDoItem"/>.
+        /// Handles request for InsertAsyncing <see cref="ToDoItem"/>.
         /// </summary>
-        /// <param name="item"><see cref="ToDoItem"/> to be inserted.</param>
+        /// <param name="item"><see cref="ToDoItem"/> to be InsertAsynced.</param>
         /// <returns>BadRequest response incase passed <see cref="ToDoItem"/> is empty or it is assigned to unexisting <see cref="ToDoList"/> or Created response with created <see cref="ToDoItem"/>.</returns>
         [HttpPost]
-        public async Task<IActionResult> AddItem([FromBody] ToDoItem item)
+        public async Task<IActionResult> AddItem([FromBody] ToDoItemCreate item)
         {
+            if ((await this.toDoListService.GetListByID(item.ToDoListID)).UserID != this.httpContextService.GetIdByContextUser(this.HttpContext.User))
+            {
+                return this.BadRequest("This list does not belong to you.");
+            }
+
             try
             {
                 var createdItem = await this.toDoItemService.AddItem(item);
@@ -161,8 +181,13 @@ namespace ToDo.WebAPI.Controllers
         /// <returns>NotFound response in case there is no such <see cref="ToDoItem"/> or Accepted response in case it was successfully deleted.</returns>
         [HttpDelete]
         [Route("{id:int}")]
-        public IActionResult DeleteItem(int id)
+        public async Task<IActionResult> DeleteItem(int id)
         {
+            if ((await this.toDoItemService.GetItem(id)).TodoList.UserID != this.httpContextService.GetIdByContextUser(this.HttpContext.User))
+            {
+                return this.BadRequest("This list does not belong to you.");
+            }
+
             try
             {
                 this.toDoItemService.DeleteItem(id);
